requirements.txt
transformers
sentence-transformers
torch
gradio
pdfplumber
python-docx
pandas
huggingface_hub